System.Posix.PAM.Types

Documentation

data PamMessage

Constructors

PamMessage

Fields

pmString :: String

pmStyle :: PamStyle

data PamStyle

Constructors

PamPromptEchoOff

PamPromptEchoOn

PamErrorMsg

PamTextInfo

data PamResponse

http://www.kernel.org/pub/linux/libs/pam/Linux-PAM-html/adg-interface-of-app-expected.html#adg-pam_conv - resp_code member in C sturct is unused and should be set to zero, that's why there is no code field in the haskell data type

Constructors

PamResponse String

data PamRetCode

Constructors

PamSuccess

PamRetCode Int

data PamFlag

Constructors

PamSilent

PamDisallowNullAuthTok

PamEstablishCred

PamDeleteCred

PamReinitializeCred

PamRefreshCred

PamChangeExpiredAuthTok

PamNone

type PamConv = Ptr () -> [PamMessage] -> IO [PamResponse]

data PamHandle

Constructors

PamHandle

Fields

cPamHandle :: Ptr ()

cPamCallback :: FunPtr (CInt -> Ptr (Ptr ()) -> Ptr (Ptr ()) -> Ptr () -> IO CInt)